+++ host/context.go
@@ -24,7 +24,7 @@ import (
 	"context"
 	"time"
 
-	"github.com/temporalio/temporal/common/client"
+	"github.com/temporalio/temporal/common/headers"
 )
 
 // NewContext create new context with default timeout 90 seconds.
@@ -35,5 +35,5 @@ func NewContext() context.Context {
 
 // NewContextWithCancel create new context with timeout.
 func NewContextWithCancel(timeout time.Duration) (context.Context, context.CancelFunc) {
-	return context.WithTimeout(client.SetHeaders(context.Background()), timeout)
+	return context.WithTimeout(headers.SetVersions(context.Background()), timeout)
 }
