+++ service/frontend/dcRedirectionPolicy.go
@@ -26,8 +26,8 @@ import (
 
 	"github.com/temporalio/temporal/.gen/go/shared"
 	"github.com/temporalio/temporal/common/cache"
-	"github.com/temporalio/temporal/common/client"
 	"github.com/temporalio/temporal/common/cluster"
+	"github.com/temporalio/temporal/common/headers"
 	"github.com/temporalio/temporal/common/service/config"
 )
 
@@ -167,7 +167,7 @@ func (policy *SelectedAPIsForwardingRedirectionPolicy) getTargetClusterAndIsDoma
 		return policy.currentClusterName, false
 	}
 
-	enforceDCRedirection := client.GetHeadersValue(ctx, client.EnforceDCRedirection)[0]
+	enforceDCRedirection := headers.GetValues(ctx, headers.EnforceDCRedirection)[0]
 	if !policy.config.EnableDomainNotActiveAutoForwarding(domainEntry.GetInfo().Name) && enforceDCRedirection != "true" {
 		// do not do dc redirection if auto-forwarding dynamic config and EnforceDCRedirection context flag is not enabled
 		return policy.currentClusterName, false
