+++ service/history/decisionHandler_test.go
@@ -23,19 +23,17 @@ package history
 import (
 	"testing"
 
-	"github.com/uber-go/tally"
-
-	"github.com/temporalio/temporal/common/log/loggerimpl"
-	"github.com/temporalio/temporal/common/metrics"
-	"github.com/temporalio/temporal/common/persistence"
-
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
+	"github.com/uber-go/tally"
 
 	"github.com/temporalio/temporal/.gen/go/shared"
 	"github.com/temporalio/temporal/common"
-	"github.com/temporalio/temporal/common/client"
+	"github.com/temporalio/temporal/common/headers"
+	"github.com/temporalio/temporal/common/log/loggerimpl"
+	"github.com/temporalio/temporal/common/metrics"
+	"github.com/temporalio/temporal/common/persistence"
 )
 
 type (
@@ -60,7 +58,7 @@ func (s *DecisionHandlerSuite) SetupTest() {
 	s.controller = gomock.NewController(s.T())
 
 	s.decisionHandler = &decisionHandlerImpl{
-		versionChecker: client.NewVersionChecker(),
+		versionChecker: headers.NewVersionChecker(),
 		metricsClient:  metrics.NewClient(tally.NoopScope, metrics.History),
 		config:         NewDynamicConfigForTest(),
 		logger:         loggerimpl.NewNopLogger(),
@@ -81,28 +79,28 @@ func (s *DecisionHandlerSuite) TearDownTest() {
 
 func (s *DecisionHandlerSuite) TestHandleBufferedQueries_ClientNotSupports() {
 	s.assertQueryCounts(s.queryRegistry, 10, 0, 0, 0)
-	s.decisionHandler.handleBufferedQueries(s.mockMutableState, client.GoSDK, "0.0.0", nil, false, testGlobalDomainEntry, false)
+	s.decisionHandler.handleBufferedQueries(s.mockMutableState, headers.GoSDK, "0.0.0", nil, false, testGlobalDomainEntry, false)
 	s.assertQueryCounts(s.queryRegistry, 0, 0, 0, 10)
 }
 
 func (s *DecisionHandlerSuite) TestHandleBufferedQueries_HeartbeatDecision() {
 	s.assertQueryCounts(s.queryRegistry, 10, 0, 0, 0)
 	queryResults := s.constructQueryResults(s.queryRegistry.getBufferedIDs()[0:5], 10)
-	s.decisionHandler.handleBufferedQueries(s.mockMutableState, client.GoSDK, client.GoWorkerConsistentQueryVersion, queryResults, false, testGlobalDomainEntry, true)
+	s.decisionHandler.handleBufferedQueries(s.mockMutableState, headers.GoSDK, headers.GoWorkerConsistentQueryVersion, queryResults, false, testGlobalDomainEntry, true)
 	s.assertQueryCounts(s.queryRegistry, 10, 0, 0, 0)
 }
 
 func (s *DecisionHandlerSuite) TestHandleBufferedQueries_NewDecisionTask() {
 	s.assertQueryCounts(s.queryRegistry, 10, 0, 0, 0)
 	queryResults := s.constructQueryResults(s.queryRegistry.getBufferedIDs()[0:5], 10)
-	s.decisionHandler.handleBufferedQueries(s.mockMutableState, client.GoSDK, client.GoWorkerConsistentQueryVersion, queryResults, true, testGlobalDomainEntry, false)
+	s.decisionHandler.handleBufferedQueries(s.mockMutableState, headers.GoSDK, headers.GoWorkerConsistentQueryVersion, queryResults, true, testGlobalDomainEntry, false)
 	s.assertQueryCounts(s.queryRegistry, 5, 5, 0, 0)
 }
 
 func (s *DecisionHandlerSuite) TestHandleBufferedQueries_NoNewDecisionTask() {
 	s.assertQueryCounts(s.queryRegistry, 10, 0, 0, 0)
 	queryResults := s.constructQueryResults(s.queryRegistry.getBufferedIDs()[0:5], 10)
-	s.decisionHandler.handleBufferedQueries(s.mockMutableState, client.GoSDK, client.GoWorkerConsistentQueryVersion, queryResults, false, testGlobalDomainEntry, false)
+	s.decisionHandler.handleBufferedQueries(s.mockMutableState, headers.GoSDK, headers.GoWorkerConsistentQueryVersion, queryResults, false, testGlobalDomainEntry, false)
 	s.assertQueryCounts(s.queryRegistry, 0, 5, 5, 0)
 }
 
@@ -114,7 +112,7 @@ func (s *DecisionHandlerSuite) TestHandleBufferedQueries_QueryTooLarge() {
 	for k, v := range largeQueryResults {
 		queryResults[k] = v
 	}
-	s.decisionHandler.handleBufferedQueries(s.mockMutableState, client.GoSDK, client.GoWorkerConsistentQueryVersion, queryResults, false, testGlobalDomainEntry, false)
+	s.decisionHandler.handleBufferedQueries(s.mockMutableState, headers.GoSDK, headers.GoWorkerConsistentQueryVersion, queryResults, false, testGlobalDomainEntry, false)
 	s.assertQueryCounts(s.queryRegistry, 0, 5, 0, 5)
 }
